HA support for ovnkube masters and for OVN DBs.
This patch adds HA support for ovnkube masters and configures
the OVN ovsdb-servers in active/standby mode so that the OVN db
contents are replicated.

With this patch, we can run multiple ovnkube in master mode.
Only one will be active. The client-go/tools/leaderelection module
is used to elect a leader.

When the ovnkube is run in the ha master mode, it participates in the
election. The ovnkube which wins the election, does the following
   - Creates or updates the endpoint - ovnkube-db with its IP address.
   - Configures the OVN ovsdb-servers as active, so that the OVN ovsdb-servers
     running in the pod run in active mode and allow write transactions.
   - Starts monitoring the pod events.

The ovnkube which loses election, does the following
   - Gets the IP address of the leader from the endpoint - ovnkube-db
   - Configures the OVN ovsdb-server to replicate the OVN DB contents
     from the leader IP. These ovsdb-servers allow only read-only transactions.
   - Stops monitoring for the pod events.

A couple of configuration options are added:
 - init-ha-master - HA is enabled if this option is set. If init-master is defined,
   then ovnkube runs in the legacy master mode.
 - ovnkube-pod-ip - The IP of pod allocated by k8s. Required when ha is enabled.

Signed-off-by: Numan Siddique <nusiddiq@redhat.com>