HA support for ovnkube masters and for OVN DBs.
This patch adds HA support for ovnkube masters and configures
the OVN ovsdb-servers in active/standby mode so that the OVN db
contents are replicated.

With this patch, we can run multiple ovnkube in master mode.
Only one will be active. The client-go/tools/leaderelection module
is used to elect a leader.

When the ovnkube is started, it participates in the
election. The ovnkube which wins the election, does the following
   - If manage-db-servers option is set, then
      * Creates or updates the endpoint - ovnkube-db with its IP address.
      * Configures the OVN ovsdb-servers as active, so that the OVN ovsdb-servers
        running in the pod allow write transactions.
   - Starts monitoring the pod events.

The ovnkube which loses election, does the following
   - If manage-db-servers option is set, then
     * Gets the IP address of the leader from the endpoint - ovnkube-db
     * Configures the OVN ovsdb-server to replicate the OVN DB contents
       from the leader IP. These ovsdb-servers allow only read-only transactions.
   - Stops monitoring for the pod events.

A couple of configuration options are added:
 - manage-db-servers (boolean) - If set, manages the OVN ovsdb-servers in active/standby
                                 mode. Uses the unix sockets to communicate with
                                 the ovsdb-servers.

 - ovnkube-pod-ip - The IP of pod allocated by k8s. Required when manage-db-servers is enabled.

There is no change required in ovnkube's running on the nodes as it monitors for
any changes to the ovnkube-db endpoint and updates the ovn-remote option so that
ovn-controller can connect to the active ovsdb-server.

A follow up patch will add the documentation.

Signed-off-by: Numan Siddique <nusiddiq@redhat.com>