ovnkube-local: Wait for all annotations before provisioning.
This is a poor man's solution.  The best way would be to handle
incremental updates to the annotations.

Also, we currently don't handle node updates at all.  We should, at
least, restart ovnkube-local when the local node changes.

This patch also trims the LocalController structure.  We don't need to
duplicate the fields that are already in the OvnController.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>