ovnkube-node: Set node zone name in the node annotation.
ovnkube-node during startup will read the Southbound database
zone name from the Southbound database with the command :
"ovn-sbctl get SB_Global . options:name" and stores it in the
node annotation - k8s.ovn.org/ovn-zone.

If the zone name is not set in the Southbound database, default
value "global" is used.  It is expected that during the deployment,
the zone name is set.

TODO: Unit test cases.

Signed-off-by: Numan Siddique <numans@ovn.org>