ovnkube-master: Manage the nodes which belong to local zone.
ovnkube-master creates the node logical switch and other Northbound
resources only if the node belongs to its zone.  Nodes with other zones
are ignored.  It also creates pod resources only if it is scheduled
in the local zone node.

ovnkube-master gets its zone name from the NB_Global.name.  If the name
is empty then the default value "global" is used.

Signed-off-by: Numan Siddique <numans@ovn.org>